Conflicts:
	app/src/main/java/com/william/mangoreader/activity/MangaItemActivity.java
	app/src/main/java/com/william/mangoreader/fragment/MangaItemChapterFragment.java
	app/src/main/java/com/william/mangoreader/model/MangaEdenMangaDetailItem.java
	app/src/main/java/com/william/mangoreader/parse/MangaEden.java

+++ app/src/main/java/com/william/mangoreader/model/MangaEdenMangaDetailItem.java
@@ -21,7 +21,7 @@ public class MangaEdenMangaDetailItem implements Serializable {
     private int numChapters;
 
     @SerializedName("created")
-    private float dateCreated;
+    private long dateCreated;
 
     @SerializedName("description")
     private String description;
